Move span factoring to helpers
